clean up scheduler metrics
